[#54]; refactor: data class를 class로 변경
