+++ src/main/kotlin/com/yourssu/signal/api/NicknameGeneratedController.kt
@@ -3,8 +3,8 @@ package com.yourssu.signal.api
 import com.yourssu.signal.config.security.annotation.RequireAuth
 import com.yourssu.signal.domain.common.business.dto.Response
 import com.yourssu.signal.api.dto.NicknameSuggestedRequest
-import com.yourssu.signal.infrastructure.dto.NicknameSuggestedResponse
-import com.yourssu.signal.infrastructure.ChatModel
+import com.yourssu.signal.infrastructure.openai.dto.NicknameSuggestedResponse
+import com.yourssu.signal.infrastructure.openai.ChatModel
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.security.SecurityRequirement
 import io.swagger.v3.oas.annotations.tags.Tag
