--- src/main/kotlin/com/yourssu/signal/infrastructure/ChatModel.kt
@@ -1,7 +1,0 @@
-package com.yourssu.signal.infrastructure
-
-import com.yourssu.signal.infrastructure.dto.NicknameSuggestedResponse
-
-interface ChatModel {
-    fun suggestNickname(statements: List<String>): NicknameSuggestedResponse
-}
