+++ src/main/kotlin/com/yourssu/signal/infrastructure/openai/ChatModel.kt
@@ -1,0 +1,7 @@
+package com.yourssu.signal.infrastructure.openai
+
+import com.yourssu.signal.infrastructure.openai.dto.NicknameSuggestedResponse
+
+interface ChatModel {
+    fun suggestNickname(statements: List<String>): NicknameSuggestedResponse
+}
