+++ src/main/kotlin/com/yourssu/signal/infrastructure/openai/dto/NicknameSuggestedResponse.kt
@@ -1,4 +1,4 @@
-package com.yourssu.signal.infrastructure.dto
+package com.yourssu.signal.infrastructure.openai.dto
 
 data class NicknameSuggestedResponse(
     val nickname: String,
