+++ src/main/kotlin/com/yourssu/signal/infrastructure/sms/exception/InvalidBankException.kt
@@ -1,4 +1,4 @@
-package com.yourssu.signal.infrastructure.deposit.exception
+package com.yourssu.signal.infrastructure.sms.exception
 
 import com.yourssu.signal.handler.BadRequestException
